convert test to TS
