--- packages/babel-preset/__fixtures__/unsupported-css-properties/code.js
@@ -1,80 +1,0 @@
-import { makeStyles } from '@griffel/react';
-
-export const useStyles = makeStyles({
-  root: {
-    all: 'value',
-    animation: 'value',
-    background: 'value',
-    backgroundPosition: 'value',
-    border: 'value',
-    borderBlock: 'value',
-    borderBlockEnd: 'value',
-    borderBlockStart: 'value',
-    borderBottom: 'value',
-    borderColor: 'value',
-    borderImage: 'value',
-    borderInline: 'value',
-    borderInlineEnd: 'value',
-    borderInlineStart: 'value',
-    borderLeft: 'value',
-    borderRadius: 'value',
-    borderRight: 'value',
-    borderStyle: 'value',
-    borderTop: 'value',
-    borderWidth: 'value',
-    columns: 'value',
-    columnRule: 'value',
-    flex: 'value',
-    flexFlow: 'value',
-    font: 'value',
-    gap: 'value',
-    grid: 'value',
-    gridArea: 'value',
-    gridColumn: 'value',
-    gridRow: 'value',
-    gridTemplate: 'value',
-    lineClamp: 'value',
-    listStyle: 'value',
-    margin: 'value',
-    mask: 'value',
-    maskBorder: 'value',
-    motion: 'value',
-    offset: 'value',
-    outline: 'value',
-    overflow: 'value',
-    overscrollBehavior: 'value',
-    padding: 'value',
-    placeItems: 'value',
-    placeSelf: 'value',
-    textDecoration: 'value',
-    textEmphasis: 'value',
-    transition: 'value',
-
-    // Vendor shorthands
-    MozAnimation: 'value',
-    MozBorderImage: 'value',
-    MozColumnRule: 'value',
-    MozColumns: 'value',
-    MozTransition: 'value',
-    WebkitAnimation: 'value',
-    WebkitBorderBefore: 'value',
-    WebkitBorderImage: 'value',
-    WebkitBorderRadius: 'value',
-    WebkitColumnRule: 'value',
-    WebkitColumns: 'value',
-    WebkitFlex: 'value',
-    WebkitFlexFlow: 'value',
-    WebkitMask: 'value',
-    WebkitMaskBoxImage: 'value',
-    WebkitTextEmphasis: 'value',
-    WebkitTextStroke: 'value',
-    WebkitTransition: 'value',
-    msContentZoomLimit: 'value',
-    msContentZoomSnap: 'value',
-    msFlex: 'value',
-    msScrollLimit: 'value',
-    msScrollSnapX: 'value',
-    msScrollSnapY: 'value',
-    msTransition: 'value',
-  },
-});
