+++ packages/babel-preset/__fixtures__/unsupported-css-properties/code.ts
@@ -1,0 +1,100 @@
+import { makeStyles } from '@griffel/react';
+
+export const useStyles = makeStyles({
+  root: {
+    // @ts-expect-error unsupported css properties should be stripped
+    all: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    animation: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    background: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    backgroundPosition: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    border: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderBlock: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderBlockEnd: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderBlockStart: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderBottom: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderColor: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderImage: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderInline: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderInlineEnd: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderInlineStart: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderLeft: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderRadius: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderRight: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderStyle: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderTop: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    borderWidth: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    columns: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    columnRule: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    flex: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    flexFlow: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    font: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    gap: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    grid: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    gridArea: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    gridColumn: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    gridRow: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    gridTemplate: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    lineClamp: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    listStyle: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    margin: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    mask: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    maskBorder: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    motion: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    offset: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    outline: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    overflow: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    overscrollBehavior: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    padding: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    placeItems: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    placeSelf: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    textDecoration: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    textEmphasis: 'value',
+    // @ts-expect-error unsupported css properties should be stripped
+    transition: 'value',
+  },
+});
